Move transformer_worker into common module
