+++ .sbt_metadata/transformer_miro.json
@@ -2,7 +2,6 @@
   "id" : "transformer_miro",
   "folder" : "pipeline/transformer/transformer_miro",
   "dependencyIds" : [
-    "internal_model",
-    "big_messaging_typesafe"
+    "transformer_common"
   ]
 }
