+++ .sbt_metadata/transformer_sierra.json
@@ -2,8 +2,7 @@
   "id" : "transformer_sierra",
   "folder" : "pipeline/transformer/transformer_sierra",
   "dependencyIds" : [
-    "internal_model",
-    "big_messaging_typesafe",
+    "transformer_common",
     "sierra_adapter_common"
   ]
 }
