+++ pipeline/transformer/Makefile
@@ -7,7 +7,7 @@ PROJECT_ID = catalogue_pipeline
 
 SBT_APPS = transformer_miro transformer_sierra transformer_mets transformer_calm
 
-SBT_DOCKER_LIBRARIES    =
+SBT_DOCKER_LIBRARIES    = transformer_common
 SBT_NO_DOCKER_LIBRARIES =
 
 PYTHON_APPS =
