+++ pipeline/transformer/transformer_common/docker-compose.yml
@@ -1,0 +1,4 @@
+sqs:
+  image: s12v/elasticmq
+  ports:
+    - "9324:9324"
