+++ pipeline/transformer/transformer_common/src/test/scala/uk/ac/wellcome/transformer/common/worker/TransformerWorkerTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.transformer.calm
+package uk.ac.wellcome.transformer.common.worker
 
 import io.circe.{Decoder, Encoder}
 import io.circe.generic.semiauto._
@@ -14,12 +14,7 @@ import uk.ac.wellcome.bigmessaging.memory.MemoryBigMessageSender
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.sqs.SQSStream
 import uk.ac.wellcome.models.Implicits._
-import uk.ac.wellcome.models.work.internal.{
-  SourceIdentifier,
-  TransformedBaseWork,
-  UnidentifiedWork,
-  WorkData
-}
+import uk.ac.wellcome.models.work.internal.{IdentifierType, SourceIdentifier, TransformedBaseWork, UnidentifiedWork, WorkData}
 import uk.ac.wellcome.monitoring.memory.MemoryMetrics
 import uk.ac.wellcome.storage.Version
 import uk.ac.wellcome.storage.maxima.Maxima
@@ -39,7 +34,7 @@ object TestTransformer extends Transformer[TestData] {
           UnidentifiedWork(
             sourceIdentifier = SourceIdentifier(
               value = "id",
-              identifierType = CalmIdentifierTypes.recordId,
+              identifierType = IdentifierType("miro-image-number"),
               ontologyType = "IdentifierType"),
             version = version,
             data = WorkData()
